requirements.txt
flask
gunicorn
groq
python-dotenv
jinja2
